Fixed warnings and APIServer under Linux.

+++ include/eve-server/admin/CommandDispatcher.h
@@ -12,7 +12,7 @@ class CommandDispatcher {
 public:
 	//this is the prototype for a command function:
 	typedef PyResult (*CommandFunc)(Client *who, CommandDB *db, PyServiceMgr *services, const Seperator &args);
-	
+
 	class CommandRecord {
 	public:
         CommandRecord( const char * cmd, const char * desc, uint32 req_role, CommandFunc& callback ) : command(cmd),
@@ -29,12 +29,12 @@ public:
 
 	PyResult Execute(Client *from, const char *msg);
 
-	void AddCommand(const char *cmd, const char *desc, uint32 required_role, CommandFunc function);
-	
+	void AddCommand(const char *cmd, const char *desc, uint64 required_role, CommandFunc function);
+
 protected:
 	PyServiceMgr &m_services;
 	CommandDB m_db;
-	
+
 	std::map<std::string, CommandRecord *> m_commands;	//we own these pointers
 };
